Removed disambigous typing by default.
To ensure proper typing I removed the whole "get entities
and components" part from `get_entities_with` and moved
back to the simpler "get entities". Additionally I changed
the example in the readme to encourage the usage of the
`get_component` over `get_components` as the later doesn't
implement proper typing as of now. Furthermore I'm not to
sure if I'll ever find a "clean" way to do so anyways...